finish clean-up for rgbaToTrle

+++ include/imgconverter.h
@@ -4,7 +4,7 @@
 void initImageConverter(int width, int height);
 void rgbaToGrayscale(uint8_t *rgba, uint8_t *gray);
 void rgbaToDxt1(uint8_t *rgba, uint8_t *dxt1);
-void rgbaToTrle(uint8_t *rgba, uint8_t *trle, uint32_t *buffer_size,uint8_t *run_offsets );
+void rgbaToTrle(uint8_t *rgba, uint8_t *trle, uint32_t *buffer_size, uint32_t *run_offsets);
 void finalizeImageConverter();
-void trleToRgba(uint8_t *rgba, uint8_t *trle, uint32_t *buffer_size,uint8_t *run_offsets);
+void trleToRgba(uint8_t *rgba, uint8_t *trle, uint32_t *buffer_size, uint32_t *run_offsets);
